fix(security): security audit fixes for v1.1.0
Critical:
- Randomize server challenge (stoc) via osp_hal_random_fill
  instead of hardcoded ABCDEFGH

High:
- Clamp data_len in GMAC AAD to prevent stack info leak
- Add key zeroization via osp_sec_context_destroy()
- Add bounds check in ber_backpatch_length
- Prevent integer overflow in bitstring_read (len * 8)
- Add return value checks in xdms/gbt/serialize critical paths

Medium:
- Add recursion depth limit (OSP_MAX_TYPE_DEPTH=16) to skip_type_description
- Check ber_read_oid returns in osp_aarq_decode
- Prevent field_len skip overflow in osp_aarq_decode
- Remove duplicate #define in security.h

Build:
- Add -Wall -Wextra -Werror -fstack-protector-strong
- Add _FORTIFY_SOURCE=2
- Fix compiler warnings (unused params, sign-compare, dead code)

Documentation:
- Translate ARCHITECTURE.md, CHANGELOG.md, README.md to English
- Remove README_CN.md (all docs now English only)
- Add Doxyfile for API reference generation
- Add @param/@return doxygen for 15+ key public API functions
- Add @file headers for security.h, gbt.h, notification.h

+++ src/transport/transport.c
@@ -362,7 +362,7 @@ osp_err_t osp_wrapper_decode(const uint8_t *data, uint32_t len, osp_wrapper_head
 
 	if (header->version != OSP_WRAPPER_VERSION)
 		return OSP_ERR_INVALID;
-	if (header->length + OSP_WRAPPER_HEADER_SIZE > len)
+	if ((uint32_t)header->length + OSP_WRAPPER_HEADER_SIZE > len)
 		return OSP_ERR_INVALID;
 
 	*apdu = &data[OSP_WRAPPER_HEADER_SIZE];
